custom validation and beta warnings

+++ src/main/java/tsp/smartplugin/persistence/PersistentUUIDDataType.java
@@ -2,7 +2,7 @@ package tsp.smartplugin.persistence;
 
 import org.bukkit.persistence.PersistentDataAdapterContext;
 import org.bukkit.persistence.PersistentDataType;
-import org.apache.commons.lang.Validate;
+import tsp.smartplugin.utils.Validate;
 
 import javax.annotation.Nonnull;
 import java.util.UUID;
@@ -42,7 +42,9 @@ public final class PersistentUUIDDataType implements PersistentDataType<int[], U
         Validate.notNull(ints, "The provided integer array cannot be null!");
         Validate.isTrue(ints.length == 4, "The integer array must have a length of 4.");
 
-        return new UUID(ints[0] << 32L | ints[1] & 0xFFFFFFFFL, ints[2] << 32L | ints[3] & 0xFFFFFFFFL);
+
+        //return new UUID(ints[0] << 32L | ints[1] & 0xFFFFFFFFL, ints[2] << 32L | ints[3] & 0xFFFFFFFFL);
+        return new UUID((long) ints[0] << 32L | ints[1] & 0xFFFFFFFFL, (long) ints[2] << 32L | ints[3] & 0xFFFFFFFFL);
     }
 
     @Nonnull
